Feat/fix code

+++ src/features/detail/ui/DetailContents.tsx
@@ -25,6 +25,7 @@ interface Props {
     userId: string
     kind: string
   }) => Promise<void>
+  noDeleteItem: boolean
 }
 
 export function DetailContents({
@@ -37,7 +38,8 @@ export function DetailContents({
   memo,
   reactions,
   onChangeArticleToggle,
-  handleReactions
+  handleReactions,
+  noDeleteItem
 }: Props) {
   const navigate = useNavigate()
   const toggleRef = useRef<HTMLDivElement>(null)
@@ -58,6 +60,10 @@ export function DetailContents({
 
   const handleDelete = async (id: string) => {
     try {
+      if (noDeleteItem) {
+        alert('할부 또는 반복 일정은 삭제할 수 없어요.')
+        return
+      }
       await deleteAccountItem(id)
       navigate(-1)
     } catch (error) {
